Created Just like Visa.
This is a description of Visa's creation I pulled together to help those working on identity systems and specifically trust frameworks think through the legacy they are invoking when they talk about Visa and its system. I think we are at the stage where thinking these things through is a good thing to do.