Updated Amira Engagement Model. Fixed Author Order
Joe Andrieu is the primary Author of the paper - he lead the group for three days straight. 
The only woman of color who contributed should be centered after that after that since this is a use case about a woman of color and well there was only one in the room.  
I took my name off the list - while I had initial interest in the paper so far I have not contributed significantly in anyway (this may change - but unlikely as I am working on my masters thesis).